Rename collection chart to filterchart
